Fix bikereg csv import
There were two problems with it. Firstly, Qt doesn't handle relations well
when it comes to using QSqlRecord to insert completely new records.

Secondly, we were shadowing a parent class's dataChanged slot with a member
function (which is all a slot is), so the parent class's dataChanged never
gets called. This parent class happened to be QTableView, so it had the
appearance of not being able to insert new rows.