Add fieldIndex() to ExtraColumnsProxyModel
This just passes the request through to the source model. Note that our
extra columns are not fields, so this is not a function for getting the
column index, given a field name. The field name specifically refers to
the field in the database model, and so we should correctly return -1 in
that case.