Fix import regression with field non-model column update
Our non-model columns in the field table view are calculated from stuff
gotten from the racer table view, so when the racer table view's model
changes, we have the dataChanged signal emit into our field table view's
updateNonModelColumns.

In updateNonModelColumns, we were in turn emitting dataChanged signal, but
for the field table view, to force a refresh of the field table view. I
think this caused some circular signal/slots thing that eventually causes
a segfault (not even a python exception!).

So instead of the big hammer of emitting that the field table model changed,
just call the field table view's dataChanged() slot directly (since the
field table model didn't really change, nobody else needs to know, so we
don't need to emit the dataChanged signal. The only thing that cares is
those non-model columns in the field table view.