Added always on top preference checkbox
This actually doesn't quite work, and seems to be a Qt5 regression (works
in Qt4). Leaving the code as is, since it should work, and maybe it'll be
fixed one day. Meanwhile, the bug is pretty benign...it just doesn't go from
on to off (off to on is okay).