Fixed field drop-down in racer view regression
Seems that any sort of proxy model messes with the relational drop-down of
the QSqlRelationalTableModel. Since the main racer table is the only place
where we make use of that, and all we need for that view is coloring, we
move the coloring from the proxy model to the actual model.