Refactor model.fieldIndex calls
Getting the column index from the database field name is so common that a
refactoring is in order. For each database table, get the column indices
ahead of time, and just use those saved-off versions in the rest of the
code.

Also, discovered we don't need the "new" argument to the table models'
create_table() method anymore.