Use milliseconds from reference datetime
For start and finish times, we want a time delta from the reference
datetime. Since Qt doesn't have a time delta class, we use milliseconds
(msecs throughout the code).

We implement an MSecsColumnsProxyModel that knows how to convert back and
forth between the milliseconds representation and some user-friendly time
representation.

Unfortunately, we can't send QDateTime objects to the QTableView, because
Qt seems completely inflexible as far as changing the datetime format that
is used to call QDateTime's toString and fromString, and its default format
is not suitable for our use. Most egregiously, it lacks milliseconds.

So, we send back strings. The user is deprived of friendly spinboxes for
editing these time delta strings.