Clean up proxy models
Right now, when proxy models are used, some instances of that table model
class use a proxy model, and some don't.  Make it consistent...if a proxy
model is sometimes used, then always use it.  Also, save a reference to the
proxy model, and to the source model, so we can refer to them directly
(instead of digging through the self.model().sourceModel() stack).