Adjust start/finish msecs in racer model
When the reference clock changes, adjust all of the start and finish msecs
values in the racer model to account for the difference in reference clock.

Visually, if you look at the racer table view, it should look like nothing
changes, because in that view, we are showing a time instead of an msecs
timedelta.