Fix bug in changing start time
When changing start time, the time in the race builder's
start_time_datetimeedit widget (a QDateTimeEdit) is used...but the date was
never initialized, so it defaulted to like Jan 1 2000. This resulted in a
start time that's way in the past, which confused some time to string
conversion routines downstream (mainly noticeable as garbled finish times
in the result reports).