Add pyinstaller build
Add build.sh script to build pyinstaller one-file executable.

The executable will be located in the out directory.

There is a workaround for pyinstaller's inability to deal with keypass's use
of the entrypoint module (located in common.py). We basically have to select
the keypass backend explicitly, instead of relying on keypass's ability to
auto-detect viable keypass backends.