ontheday: Use checksums to update changes
When website or other clients make changes, periodically check for those
changes and update our client accordingly. We use the various checksums
available in the ontheday json records to detect changes. These changes
are performed in a separate worker thread, and the changes are collated for
processing by the main thread.