Fix bug with field start time
When working on the Calaveras TT, some code was added that broke the case
where an entire field has the same start time (i.e. the field has a
"time_start" field). The broken code was the new code that picked up changes
from ontheday.net (to support new racer registration during the event, as
well as multiple clients).